% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AC_IPD_binY_contX}
\alias{AC_IPD_binY_contX}
\title{Individual-level patient data for binary outcome, continuous covariates}
\format{
\subsection{\code{y ~ PF1 + PF2 + trt + trt:(EM1 + EM2)}}{

\describe{
\item{id}{Numeric unique identifier}
\item{PF_cont_1}{Numeric prognostic factor continuous covariate}
\item{PF_cont_2}{Numeric prognostic factor continuous covariate}
\item{EM_cont_1}{Numeric effect modifier continuous covariate}
\item{EM_cont_2}{Numeric effect modifier continuous covariate}
\item{trt}{Factor treatment identifier. Levels A, C}
\item{y}{Integer binary outcome}
\item{true_eta}{Numeric linear predictor}
}
}
}
\source{
Simulated data
}
\usage{
AC_IPD_binY_contX
}
\description{
This data set contains simulated patient covariate and outcome values.
}
\references{
Remiro‐Azocar A, Heath A, Baio G (2022)
}
\keyword{datasets}
