% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AC_IPD_countY_contX}
\alias{AC_IPD_countY_contX}
\title{Individual-level patient data for count outcome, continuous covariates}
\format{
\subsection{\code{y ~ PF_cont_1 + PF_cont_2 + trt + trt:(EM_cont_1 + EM_cont_2)}}{

\describe{
\item{id}{Numeric unique identifier}
\item{PF_cont_1}{Numeric prognostic factor continuous covariate}
\item{PF_cont_2}{Numeric prognostic factor continuous covariate}
\item{EM_cont_1}{Numeric effect modifier continuous covariate}
\item{EM_cont_2}{Numeric effect modifier continuous covariate}
\item{trt}{Factor treatment identifier. Levels A, C}
\item{y}{Integer non-negative count outcome}
\item{true_eta}{Numeric linear predictor}
}
}
}
\source{
Simulated data
}
\usage{
data(AC_IPD_countY_contX)
}
\description{
This data set contains simulated patient covariate and outcome values.
Corresponds to ALD data set.
}
\references{
Remiro‐Azocar A, Heath A, Baio G (2022)
}
\keyword{datasets}
