% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{ALD_stats}
\alias{ALD_stats}
\title{Aggregate-level data statistics}
\usage{
ALD_stats(data, trials = list("B", "C"))
}
\arguments{
\item{data}{Data}

\item{trials}{Trial labels; default \code{B}, \code{C}}
}
\value{
List of marginal treatment effect and variance
}
\description{
Aggregate-level data statistics
}
