% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_ALD}
\alias{BC_ALD}
\title{Data set for aggregate level patient data in wide format}
\format{
A data list including the variables needed for a outstandR basic example. The variables are as follows:
\describe{
\item{mean.X1}{Numeric value}
\item{mean.X2}{Numeric value}
\item{mean.X3}{Numeric value}
\item{mean.X4}{Numeric value}
\item{sd.X1}{Numeric standard deviation value}
\item{sd.X2}{Numeric standard deviation value}
\item{sd.X3}{Numeric standard deviation value}
\item{sd.X4}{Numeric standard deviation value}
\item{y.B.sum}{Integer number of patients with event on B treatment}
\item{y.B.bar}{Numeric proportion of patients on B treatment}
\item{N.B}{Integer Total number of patients on B treatment}
\item{y.C.sum}{Integer number of patients with event on C treatment}
\item{y.C.bar}{Numeric proportion of patients on C treatment}
\item{N.C}{Integer total number of patients on C treatment}
}
}
\source{
Simulated data
}
\description{
This data set contains summaries of simulated patient covariate and outcome values.
}
\references{
Remiro‐Azócar A, Heath A, Baio G (2022)
}
\keyword{datasets}
