% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_ALD_contY_mixedX}
\alias{BC_ALD_contY_mixedX}
\title{Aggregate level patient data for continuous outcome, mixed covariates}
\format{
\subsection{\code{y ~ X1 + X3 + X4 + trt + trt:(X2 + X3 + X4)}}{

\describe{
\item{variable}{String covariate or outcome name. From X1, X2, X3, X4, y.}
\item{statistic}{String summary statistic name. From mean, sd, prob, sum, N}
\item{value}{Numeric value}
\item{trt}{Treatment (arm) name. From B, C}
}
}
}
\source{
Simulated data
}
\usage{
data(BC_ALD_contY_mixedX)
}
\description{
This data set contains summaries of simulated patient covariate and outcome values.
Corresponds to IPD data set.
}
\references{
Remiro‐Azocar A, Heath A, Baio G (2022)
}
\keyword{datasets}
