% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_ALD_countY_contX}
\alias{BC_ALD_countY_contX}
\title{Aggregate level patient data for count outcome, continuous covariates}
\format{
\subsection{\code{y ~ PF_cont_1 + PF_cont_2 + trt + trt:(EM_cont_1 + EM_cont_2)}}{

\describe{
\item{variable}{String covariate or outcome name. From EM_cont_1, EM_cont_2, PF_cont_1, PF_cont_2, y.}
\item{statistic}{String summary statistic name. From mean, sd, sum, N}
\item{value}{Numeric value}
\item{trt}{Treatment (arm) name. From B, C}
}
}
}
\source{
Simulated data
}
\usage{
data(BC_ALD_countY_contX)
}
\description{
This data set contains summaries of simulated patient covariate and outcome values.
Corresponds to IPD data set.
}
\references{
Remiro‐Azocar A, Heath A, Baio G (2022)
}
\keyword{datasets}
