% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_ALD_long}
\alias{BC_ALD_long}
\title{Data set for aggregate level patient data in long format}
\format{
A data frame with 4 variables:
\describe{
\item{statistic}{A character string}
\item{trt}{A factor}
\item{value}{A numeric value}
\item{variable}{A character string}
}
}
\source{
Simulated data
}
\description{
This data set contains summaries of simulated patient covariate and outcome values.
}
\keyword{datasets}
