% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{bias.mcse}
\alias{bias.mcse}
\title{Monte Carlo SE of bias estimate}
\usage{
bias.mcse(theta.hat)
}
\arguments{
\item{theta.hat}{theta hat}
}
\value{
\eqn{sqrt(1/(nsim*(nsim-1))*tmp)}
}
\description{
Monte Carlo SE of bias estimate
}
