% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_stan.R
\name{calc_gcomp_stan}
\alias{calc_gcomp_stan}
\title{G-computation using Stan}
\usage{
calc_gcomp_stan(strategy, ipd, ald)
}
\arguments{
\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{formula}{Linear regression \code{formula} object}

\item{family}{A \code{family} object}
}
\value{
A list of \eqn{y^*_A} and \eqn{y^*_C} posterior predictions
}
\description{
Calculate draws of binary responses from posterior predictive distribution
from the Bayesian G-computation method using Hamiltonian Monte Carlo.
}
