% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{calc_maic}
\alias{calc_maic}
\title{Calculate MAIC}
\usage{
calc_maic(strategy, analysis_params)
}
\arguments{
\item{strategy}{An object of class \code{strategy} created by functions such as
\code{\link[=strategy_maic]{strategy_maic()}}, \code{\link[=strategy_stc]{strategy_stc()}}, or \code{\link[=strategy_mim]{strategy_mim()}}.
Contains modelling details like the formula and family.}

\item{analysis_params}{Analysis parameters; list}
}
\value{
List of numeric treatment means.
}
\description{
Calculate MAIC
}
\keyword{internal}
