% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stc.R
\name{calc_stc}
\alias{calc_stc}
\title{Calculate simulated treatment comparison statistics}
\usage{
calc_stc(strategy, analysis_params, ...)
}
\arguments{
\item{strategy}{An object of class \code{strategy} created by functions such as
\code{\link[=strategy_maic]{strategy_maic()}}, \code{\link[=strategy_stc]{strategy_stc()}}, or \code{\link[=strategy_mim]{strategy_mim()}}.
Contains modelling details like the formula and family.}

\item{analysis_params}{List of analysis parameters}

\item{...}{Additional arguments}
}
\value{
A list:
\describe{
\item{\code{mean_A}}{Mean for comparator treatment group "A".}
\item{\code{mean_C}}{Mean for reference treatment group "C".}
}
}
\description{
Calculate simulated treatment comparison statistics
}
\keyword{internal}
