% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_ate}
\alias{calculate_ate}
\title{Calculate ATE}
\usage{
calculate_ate(mean_A, mean_C, family)
}
\arguments{
\item{family}{family object of the model}

\item{ppv}{model prediction samples}
}
\value{
ATE
}
\description{
Calculate ATE
}
