% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean_count}
\alias{calculate_trial_mean_count}
\title{Calculate trial mean count}
\usage{
calculate_trial_mean_count(ald, tid, effect)
}
\arguments{
\item{ald}{Aggregate level data}

\item{tid}{Treatment ID}

\item{effect}{Effect}
}
\value{
Value
}
\description{
Calculate trial mean count
}
