% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_variance_binary}
\alias{calculate_trial_variance_binary}
\title{Calculate trial variance binary}
\usage{
calculate_trial_variance_binary(ald, tid, effect)
}
\arguments{
\item{ald}{Aggregate level data}

\item{tid}{Treatment ID}

\item{effect}{Effect}
}
\value{
Numeric value of total variance.
}
\description{
Calculate trial variance binary
}
