% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_formula.R
\name{check_balance_formula}
\alias{check_balance_formula}
\title{Check balance formula}
\usage{
check_balance_formula(formula, trt_var = NULL)
}
\arguments{
\item{formula}{Formula for the balance model}

\item{trt_var}{Treatment variable (used to warn if mistakenly included)}
}
\value{
No return value, called for side effects
}
\description{
Check balance formula
}
\keyword{internal}
