% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_formula.R
\name{check_formula}
\alias{check_formula}
\title{Check outcome formula}
\usage{
check_formula(formula, trt_var = NULL)
}
\arguments{
\item{formula}{Formula for the outcome model}

\item{trt_var}{Treatment variable}
}
\value{
No return value, called for side effects
}
\description{
Check outcome formula
}
\keyword{internal}
