% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_formula.R
\name{check_formula_base}
\alias{check_formula_base}
\title{Base formula validator}
\usage{
check_formula_base(
  formula,
  trt_var = NULL,
  must_have_response = TRUE,
  must_have_trt = TRUE
)
}
\description{
Base formula validator
}
\keyword{internal}
