% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{continuity_correction}
\alias{continuity_correction}
\title{Continuity Correction}
\usage{
continuity_correction(ald, correction = 0.5)
}
\arguments{
\item{ald}{Aggregate level data}

\item{correction}{Correction size; numeric}
}
\value{
Corrected aggregate level data
}
\description{
Continuity Correction
}
\keyword{internal}
