% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Compute covariance matrix}
\usage{
cor2cov(R, S)
}
\arguments{
\item{R}{correlation matrix}

\item{S}{vector of standard deviations covariance matrix}
}
\value{
Required input for \code{mvrnorm}.
}
\description{
Compute covariance matrix
}
\keyword{internal}
