% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{coverage.mcse}
\alias{coverage.mcse}
\title{Monte Carlo SE of coverage estimate}
\usage{
coverage.mcse(coverage, nsim)
}
\arguments{
\item{coverage}{Coverage}

\item{nsim}{Number of simulations}
}
\value{
\eqn{sqrt((coverage*(1 - coverage))/nsim)}
}
\description{
Monte Carlo SE of coverage estimate
}
\keyword{internal}
