% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{coverage}
\alias{coverage}
\title{Coverage estimate}
\usage{
coverage(low, upp, theta)
}
\arguments{
\item{low}{Low}

\item{upp}{Upper}

\item{theta}{Theta}
}
\value{
\eqn{sum(in_range)/nsim}
}
\description{
Coverage estimate
}
\keyword{internal}
