% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{empse.mcse}
\alias{empse.mcse}
\title{EmpSE MCSE}
\usage{
empse.mcse(empse, nsim)
}
\arguments{
\item{empse}{EMPSE}

\item{nsim}{Number of simulations}
}
\value{
\eqn{empse/(sqrt(2*(nsim-1)))}
}
\description{
EmpSE MCSE
}
