% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{empse}
\alias{empse}
\title{Empirical standard error}
\usage{
empse(theta.hat)
}
\arguments{
\item{theta.hat}{Theta}
}
\value{
\eqn{sqrt(tmp/(nsim-1))}
}
\description{
Empirical standard error
}
\keyword{internal}
