% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-functions.R
\name{find_conversion_path}
\alias{find_conversion_path}
\title{recursive function to find conversion path}
\usage{
find_conversion_path(from, to, visited = c())
}
\description{
recursive function to find conversion path
}
\examples{
find_conversion_path("log_odds", "relative_risk")

}
