% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml_means}
\alias{gcomp_ml_means}
\title{G-computation Maximum Likelihood mean outcome}
\usage{
gcomp_ml_means(formula, family, N = 1000, ipd, ald)
}
\arguments{
\item{formula}{Linear regression \code{formula} object}

\item{family}{Family object}

\item{N}{Synthetic sample size for g-computation}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}
}
\value{
Difference on relative treatment effect scale
}
\description{
G-computation Maximum Likelihood mean outcome
}
\section{Log-Odds Ratio}{

Marginal \emph{A} vs \emph{C} log-odds ratio (mean difference in expected log-odds)
estimated by transforming from probability to linear predictor scale.

\eqn{\log(\hat{\mu}_A/(1 - \hat{\mu}_A)) - \log(\hat{\mu}_C/(1 - \hat{\mu}_C))}
}

\seealso{
\code{\link[=strategy_gcomp_ml]{strategy_gcomp_ml()}}, \code{\link[=gcomp_ml.boot]{gcomp_ml.boot()}}
}
\keyword{internal}
