% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_stan.R
\name{gcomp_stan}
\alias{gcomp_stan}
\title{G-computation using Stan}
\usage{
gcomp_stan(formula = NULL, ipd, ald)
}
\arguments{
\item{formula}{Linear regression \code{formula} object}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}
}
\value{
A list of \eqn{y^*_A} and \eqn{y^*_C} posterior predictions
}
\description{
Calculate draws of binary responses from posterior predictive distribution
from the Bayesian G-computation method using Hamiltonian Monte Carlo.
}
\keyword{internal}
