% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{generate simulated datasets for index and comparator trials}
\usage{
gen_data(N, b_trt, b_X, b_EM, b_0, meanX, sdX, event_rate, corX, allocation)
}
\description{
generate simulated datasets for index and comparator trials
}
