% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_method.R
\name{get_allowed_var_methods}
\alias{get_allowed_var_methods}
\title{Retrieve list of allowed variance methods}
\usage{
get_allowed_var_methods(strategy)
}
\arguments{
\item{strategy}{Strategy object}
}
\value{
Character vector of allowed methods. First element is the default.
}
\description{
Retrieve list of allowed variance methods
}
\keyword{internal}
