% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{get_comparator}
\alias{get_comparator}
\title{Get study comparator treatment names}
\usage{
get_comparator(dat, ref_trt, trt_var = "trt")
}
\arguments{
\item{dat}{Data frame.}

\item{ref_trt}{Reference treatment. String.}

\item{trt_var}{Treatment variable. String default "trt".}
}
\value{
Comparator string names
}
\description{
Get study comparator treatment names
}
\keyword{internal}
