% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_covariate_names}
\alias{get_covariate_names}
\title{Get covariate names}
\usage{
get_covariate_names(formula)
}
\arguments{
\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt + trt:X2. For covariates as both PF and EM use \code{*} syntax.}
}
\value{
covariate names vector
}
\description{
Get covariate names
}
\keyword{internal}
