% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_eff_mod_names}
\alias{get_eff_mod_names}
\title{Get effect modifiers}
\usage{
get_eff_mod_names(formula)
}
\arguments{
\item{formula}{Linear regression formula string}
}
\value{
Effect modifiers names
}
\description{
Get effect modifiers
}
\keyword{internal}
