% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_effect_modifiers}
\alias{get_effect_modifiers}
\title{Get effect modifiers}
\usage{
get_effect_modifiers(formula)
}
\arguments{
\item{formula}{Linear regression formula}
}
\value{
Effect modifiers names
}
\description{
Get effect modifiers
}
\keyword{internal}
