% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_mean_names}
\alias{get_mean_names}
\title{Get mean names}
\usage{
get_mean_names(ald, var_names)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{var_names}{Variable names character vector}
}
\value{
Mean names vector
}
\description{
Get mean names
}
\keyword{internal}
