% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var_sandwich.R
\name{get_robust_vcov}
\alias{get_robust_vcov}
\title{Compute Robust Covariance Matrix (HC0-style)}
\usage{
get_robust_vcov(fit)
}
\description{
Calculates (X'WX)^-1 (X' W^2 r^2 X) (X'WX)^-1
}
\keyword{internal}
