% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_treatment_name}
\alias{get_treatment_name}
\title{Get treatment name}
\usage{
get_treatment_name(formula, trt_var = NULL)
}
\arguments{
\item{formula}{Formula}

\item{trt_var}{Treatment variable}
}
\value{
Treatment name string.
}
\description{
Get treatment name
}
\keyword{internal}
