% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_method.R
\name{get_var_method}
\alias{get_var_method}
\title{Determine and validate variance method for a strategy}
\usage{
get_var_method(strategy, user_method = NULL)
}
\arguments{
\item{strategy}{Strategy object}

\item{user_method}{User supplied method (string) or NULL}
}
\value{
String name of variance method
}
\description{
Determine and validate variance method for a strategy
}
\keyword{internal}
