% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{guess_treatment_name}
\alias{guess_treatment_name}
\title{Guess treatment name}
\usage{
guess_treatment_name(formula)
}
\arguments{
\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt + trt:X2. For covariates as both PF and EM use \code{*} syntax.}
}
\value{
Treatment name
}
\description{
Does a variable appear more than once in interactions?
If not then pick first LHS interaction term.
Finally, if there are no interactions then pick last main effect term.
}
\keyword{internal}
