% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_IPD_stats.R
\name{IPD_stat_factory}
\alias{IPD_stat_factory}
\title{Factory function for creating calc_IPD_stats methods}
\usage{
IPD_stat_factory(ipd_fun)
}
\arguments{
\item{ipd_fun}{A function that computes mean and variance statistics for individual-level patient data.}
}
\value{
A function that computes mean and variance statistics for a given strategy.
}
\description{
Creates a method for computing mean and variance statistics based on the supplied function.
}
\keyword{internal}
