% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats.gcomp_ml}
\alias{IPD_stats.gcomp_ml}
\title{IPD_stats.gcomp_ml}
\usage{
\method{IPD_stats}{gcomp_ml}(strategy, data = AC.IPD)
}
\arguments{
\item{strategy}{}

\item{data}{}
}
\description{
IPD_stats.gcomp_ml
}
