% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats.gcomp_stan}
\alias{IPD_stats.gcomp_stan}
\title{IPD_stats.gcomp_stan}
\usage{
\method{IPD_stats}{gcomp_stan}(strategy, data)
}
\arguments{
\item{strategy}{}

\item{data}{}
}
\description{
IPD_stats.gcomp_stan
}
