% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats.maic}
\alias{IPD_stats.maic}
\title{IPD_stats.maic}
\usage{
\method{IPD_stats}{maic}(strategy, data = AC.IPD)
}
\arguments{
\item{strategy}{}

\item{data}{}
}
\description{
marginal A vs C treatment effect estimates
using bootstrapping
}
