% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats}
\alias{IPD_stats}
\alias{IPD_stats.default}
\alias{IPD_stats.maic}
\alias{IPD_stats.stc}
\alias{IPD_stats.gcomp_ml}
\alias{IPD_stats.gcomp_stan}
\title{Individual level data statistics}
\usage{
IPD_stats(strategy, ipd, ald, ...)

\method{IPD_stats}{default}()

\method{IPD_stats}{maic}(strategy, ipd, ald)

\method{IPD_stats}{stc}(strategy, ipd, ald)

\method{IPD_stats}{gcomp_ml}(strategy, ipd, ald)

\method{IPD_stats}{gcomp_stan}(strategy, ipd, ald)
}
\arguments{
\item{strategy}{A list corresponding to different approaches}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{...}{Additional arguments}
}
\value{
Mean and variance
}
\description{
Marginal A vs C treatment effect estimates
using bootstrapping

IPD from the AC trial are used to fit a regression model describing the
observed outcomes \code{y} in terms of the relevant baseline characteristics \code{x} and
the treatment variable \code{z}.
}
