% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{mae}
\alias{mae}
\title{Mean absolute error estimate}
\usage{
mae(theta.hat, theta)
}
\arguments{
\item{theta.hat}{Theta hat}

\item{theta}{Theta}
}
\value{
\eqn{sum(abs(theta.hat - theta))/nsim}
}
\description{
Mean absolute error estimate
}
\keyword{internal}
