% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic_weights}
\alias{maic_weights}
\title{Estimate MAIC weights}
\usage{
maic_weights(X.EM)
}
\arguments{
\item{X.EM}{Centred S=1 effect modifiers; matrix or data frame}
}
\value{
Estimated weights for each individual; vector
}
\description{
\insertCite{Signorovitch2010}{mimR}
}
\references{
\insertRef{Signorovitch2010}{mimR}
}
