% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{mcse.estimate}
\alias{mcse.estimate}
\title{Monte Carlo SE of any continuous performance metric}
\usage{
mcse.estimate(pm)
}
\arguments{
\item{pm}{pm}
}
\value{
\eqn{sqrt(sum((pm - pm_mean)^2)/(nsim*(nsim-1)))}
}
\description{
Monte Carlo SE of any continuous performance metric
}
\keyword{internal}
