% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{mim}
\alias{mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
mim(formula, ipd, ald, M = 1000, n.chains = 2, warmup = 1000, iters = 4000)
}
\arguments{
\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{M}{Number of syntheses used in analysis stage (high for low Monte Carlo error)}
}
\description{
Multiple imputation marginalization (MIM)
}
\keyword{internal}
