% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimR.R
\name{mimR}
\alias{mimR}
\title{Calculate the difference between treatments using all evidence}
\usage{
mimR(AC.IPD, BC.ALD, strategy, CI = 0.95, ...)
}
\arguments{
\item{AC.IPD}{Individual-level patient data. Suppose between studies \emph{A} and \emph{C}.}

\item{BC.ALD}{Aggregate-level data. Suppose between studies \emph{B} and \emph{C}.}

\item{strategy}{Computation strategy function. These can be
\code{strategy_maic()}, \code{strategy_stc()}, \code{strategy_gcomp_ml()} and  \code{strategy_gcomp_stan()}}

\item{CI}{Confidence interval; between 0,1}

\item{...}{Additional arguments}
}
\value{
List of length 3 of statistics as a \code{mimR} class object.
Containing statistics between each pair of treatments.
These are the mean contrasts, variances and confidence intervals,
respectively.
}
\description{
This is the main, top-level wrapper for \code{{mimR}}.
Methods taken from
\insertCite{RemiroAzocar2022}{mimR}.
}
\examples{

data(AC_IPD)  # AC patient-level data
data(BC_ALD)  # BC aggregate-level data

# matching-adjusted indirect comparison
mimR_maic <- mimR(AC_IPD, BC_ALD, strategy = strategy_maic())

# simulated treatment comparison
mimR_stc <- mimR(AC_IPD, BC_ALD, strategy = strategy_stc())

# G-computation with maximum likelihood
# mimR_gcomp_ml <- mimR(AC_IPD, BC_ALD, strategy = strategy_gcomp_ml())

# G-computation with Bayesian inference
mimR_gcomp_stan <- mimR(AC_IPD, BC_ALD, strategy = strategy_gcomp_stan())

}
\references{
\insertRef{RemiroAzocar2022}{mimR}
}
