% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{mse.mcse}
\alias{mse.mcse}
\title{Monte Carlo SE of MSE estimate}
\usage{
mse.mcse(theta.hat, theta)
}
\arguments{
\item{theta.hat}{Theta hat}

\item{theta}{Theta}
}
\value{
\eqn{sqrt(sum((tmp - mse.est)^2)/(nsim*(nsim-1)))}
}
\description{
Monte Carlo SE of MSE estimate
}
