% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{mse}
\alias{mse}
\title{Mean squared error estimate}
\usage{
mse(theta.hat, theta)
}
\arguments{
\item{theta.hat}{Theta hat}

\item{theta}{Theta}
}
\value{
\eqn{sum((theta.hat - theta)^2)/nsim}
}
\description{
Mean squared error estimate
}
