% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{outstandR-package}
\alias{outstandR-package}
\title{outstandR: Model-Based Standardisation for Indirect Treatment Comparison with Limited Subject-Level Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

For the problem of indirect treatment comparison with limited subject-level data, this package provides tools for model-based standardisation with several different computation approaches. See Remiro‐Azócar A, Heath A, Baio G (2022. Parametric G‐computation for compatible indirect treatment comparisons with limited individual patient data, Res. Synth. Methods, 1–31. ISSN 1759-2879, \doi{10.1002/jrsm.1565}.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://StatisticsHealthEconomics.github.io/outstandR/}
  \item Report bugs at \url{https://github.com/StatisticsHealthEconomics/outstandR/issues/}
}

}
\author{
\strong{Maintainer}: Nathan Green \email{n.green@ucl.ac.uk} (\href{https://orcid.org/0000-0003-2745-1736}{ORCID}) (\href{https://ror.org/02jx3x895}{ROR}) [copyright holder]

Authors:
\itemize{
  \item Chengyang Gao \email{chengyang.gao.15@ucl.ac.uk}
  \item Antonio Remiro-Azocar \email{antonio.remiro-azocar@bayer.com} (\href{https://orcid.org/0000-0002-7263-4251}{ORCID})
}

}
\keyword{internal}
