% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outstandR.R
\name{outstandR}
\alias{outstandR}
\title{Calculate the difference between treatments using all evidence}
\usage{
outstandR(AC.IPD, BC.ALD, strategy, CI = 0.95, ...)
}
\arguments{
\item{AC.IPD}{Individual-level patient data. Suppose between studies \emph{A} and \emph{C}.}

\item{BC.ALD}{Aggregate-level data. Suppose between studies \emph{B} and \emph{C}.}

\item{strategy}{Computation strategy function. These can be
\code{strategy_maic()}, \code{strategy_stc()}, \code{strategy_gcomp_ml()} and  \code{strategy_gcomp_stan()}}

\item{CI}{Confidence interval; between 0,1}

\item{...}{Additional arguments}
}
\value{
List of length 3 of statistics as a \code{outstandR} class object.
Containing statistics between each pair of treatments.
These are the mean contrasts, variances and confidence intervals,
respectively.
}
\description{
This is the main, top-level wrapper for \code{{outstandR}}.
Methods taken from
\insertCite{RemiroAzocar2022}{outstandR}.
}
\examples{
data(AC_IPD)  # AC patient-level data
data(BC_ALD)  # BC aggregate-level data

lin_form <- as.formula("y ~ X3 + X4 + trt*X1 + trt*X2")

# matching-adjusted indirect comparison
outstandR_maic <- outstandR(AC_IPD, BC_ALD, strategy = strategy_maic(formula = lin_form))

# simulated treatment comparison
outstandR_stc <- outstandR(AC_IPD, BC_ALD, strategy = strategy_stc(lin_form))

# G-computation with maximum likelihood
# outstandR_gcomp_ml <- outstandR(AC_IPD, BC_ALD, strategy = strategy_gcomp_ml(lin_form))

# G-computation with Bayesian inference
outstandR_gcomp_stan <- outstandR(AC_IPD, BC_ALD, strategy = strategy_gcomp_stan(lin_form))

# Multiple imputation marginalization
outstandR_gcomp_stan <- outstandR(AC_IPD, BC_ALD, strategy = strategy_mim(lin_form))

}
\references{
\insertRef{RemiroAzocar2022}{outstandR}
}
