% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_ald}
\alias{prep_ald}
\title{Prepare ALD}
\usage{
prep_ald(form, data, trt_var = "trt")
}
\arguments{
\item{form}{Formula}

\item{data}{A dataframe of ALD}

\item{trt_var}{Treatment variable name}
}
\description{
Prepare ALD
}
