% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_ipd}
\alias{prep_ipd}
\title{Prepare Individual Patient Data}
\usage{
prep_ipd(form, data)
}
\arguments{
\item{form}{Formula}

\item{data}{Individual patient data}
}
\value{
Model frame
}
\description{
Prepare Individual Patient Data
}
\keyword{internal}
