% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prepare_covariate_distns}
\alias{prepare_covariate_distns}
\title{Prepare Covariate Distributions}
\usage{
prepare_covariate_distns(
  formula,
  ald,
  trt_var,
  marginal_distns,
  marginal_params,
  verbose = FALSE
)
}
\description{
Resolves missing distributions and parameters by looking at the ALD.
Allows for partial specification (e.g., user specifies one variable,
function auto-detects the rest).
}
\keyword{internal}
