% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.outstandR}
\alias{print.outstandR}
\title{Print a Summary of a outstandR Object}
\usage{
\method{print}{outstandR}(x, ...)
}
\arguments{
\item{x}{Objects of the class "outstandR"}

\item{...}{Additional arguments passed to other methods}
}
\value{
No return value, called for side effects
}
\description{
This is a method for the function \code{\link[=print]{print()}} for objects of the
class "outstandR" created by a call to \code{\link[=outstandR]{outstandR()}}
}
\seealso{
\code{\link[=outstandR]{outstandR()}}
}
