% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metrics.R
\name{process_metrics}
\alias{process_metrics}
\title{Compute performance metrics for a given method}
\usage{
process_metrics(means, variances, truth)
}
\arguments{
\item{means}{Mean}

\item{variances}{Variances}

\item{truth}{Truth}
}
\description{
Compute performance metrics for a given method
}
