% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{Q}
\alias{Q}
\title{objective function to be minimized for standard method of moments}
\usage{
Q(alpha, X.EM)
}
\description{
objective function to be minimized for standard method of moments
}
