% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_ald.R
\name{reshape_ald_to_wide}
\alias{reshape_ald_to_wide}
\title{Convert aggregate data from long to wide format}
\usage{
reshape_ald_to_wide(df)
}
\arguments{
\item{df}{A Dataframe of ALD}
}
\value{
Data frame in wide format
}
\description{
Convert aggregate data from long to wide format
}
\examples{
df <- 
  data.frame(
    variable = c("age", "age", "y", "y", "y", "y", "y", "y", "y", "y"),
    statistic = c("mean", "sd", "sum", "bar", "sd", "N", "sum", "bar", "sd", "N"),
    trt = c(NA, NA, "B", "B", "B", "B", "C", "C", "C", "C"),
    value = c(1,1,1,1,1,1,1,1,1,1))

}
