% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{simulate_ALD_pseudo_pop}
\alias{simulate_ALD_pseudo_pop}
\title{Simulate Aggregate-Level Data Pseudo-Population}
\usage{
simulate_ALD_pseudo_pop(formula, ipd, ald, trt_var, rho = NA, N = 1000)
}
\arguments{
\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt + trt:X2. For covariates as both PF and EM use \code{*} syntax.}

\item{ipd}{Individual-level patient data. Dataframe with one row per patient with outcome, treatment and covariate columns.}

\item{ald}{Aggregate-level data. Single row matrix with summary statistics for each covariate and treatment outcomes.
The format is 'mean.\emph{' and 'sd.}' for covariates and 'y.\emph{.sum', 'y.}.bar', 'y.*.sd' for treatments B and C.
We assume a common distribution for each treatment arm.}

\item{rho}{A named square matrix of covariate correlations; default NA.}

\item{N}{Sample size for the synthetic cohort. Default is 1000.}
}
\value{
A data frame representing the synthetic pseudo-population.
}
\description{
Generates a synthetic cohort using a normal copula based on aggregate-level data.
}
\examples{
\dontrun{
formula <- outcome ~ treatment + age
ipd <- data.frame(treatment = c(0, 1), outcome = c(1, 0), age = c(30, 40))
ald <- data.frame()
simulate_ALD_pseudo_pop(formula, ipd, ald, N = 1000)
}
}
\keyword{internal}
