% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{strategy_gcomp_ml}
\alias{strategy_gcomp_ml}
\title{strategy_gcomp_ml}
\usage{
strategy_gcomp_ml(
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"),
  R = 1000
)
}
\arguments{
\item{formula}{}

\item{R}{}
}
\description{
strategy_gcomp_ml
}
