% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{strategy_gcomp_stan}
\alias{strategy_gcomp_stan}
\title{strategy_gcomp_stan}
\usage{
strategy_gcomp_stan(formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"))
}
\arguments{
\item{formula}{}
}
\description{
strategy_gcomp_stan
}
