% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{strategy_maic}
\alias{strategy_maic}
\title{strategy_maic}
\usage{
strategy_maic(
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"),
  R = 1000,
  dat_ALD = BC.ALD
)
}
\arguments{
\item{formula}{}

\item{R}{}

\item{dat_ALD}{}
}
\description{
strategy_maic
}
