% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_.R
\name{strategy_stc}
\alias{strategy_stc}
\title{Simulated treatment comparison (STC)}
\usage{
strategy_stc(
  formula = NULL,
  family = gaussian(link = "identity"),
  trt_var = NULL
)
}
\arguments{
\item{formula}{Model formula.}

\item{family}{Model family.}

\item{trt_var}{Treatment variable name.}
}
\value{
\code{stc} class object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{strategy_stc()} was deprecated in outstandR version 1.X.X.
We recommend using G-computation (\code{strategy_gcomp_ml()}) as a more robust
alternative for this type of analysis.

Outcome regression-based method which targets a conditional treatment effect.
STC is a modification of the covariate adjustment method.
An outcome model is fitted using IPD in the \emph{AB} trial. For example,

\deqn{
g(\mu_{t(AB)}(X)) = \beta_0 + \beta_1^T X + (\beta_B + \beta_2^T X^{EM}) I(t=B)
}
where \eqn{\beta_0} is an intercept term, \eqn{\beta_1} is a vector of coefficients for
prognostic variables, \eqn{\beta_B} is the relative effect of treatment \emph{B} compared
to \emph{A} at \eqn{X=0}, \eqn{\beta_2} is a vector of coefficients for effect
modifiers \eqn{X^{EM}} subvector of the full covariate vector \eqn{X}), and
\eqn{\mu_{t(AB)}(X)} is the expected outcome of an individual assigned
treatment \eqn{t} with covariate values \eqn{X} which is transformed onto a
chosen linear predictor scale with link function \eqn{g(\cdot)}.
}
