% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.outstandR}
\alias{summary.outstandR}
\title{Summary method for outstandR}
\usage{
\method{summary}{outstandR}(object, ...)
}
\arguments{
\item{object}{\code{\link[=outstandR]{outstandR()}} output object}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects
}
\description{
Summary method for outstandR
}
