% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALD_stats.R
\name{trial_treatment_effect}
\alias{trial_treatment_effect}
\title{Trial treatment effect with aggregate data}
\usage{
trial_treatment_effect(ald, tid, family)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{tid}{Treatment label}

\item{family}{}
}
\value{
Value
}
\description{
Calculate
\deqn{\log(\sum y_k (N_k - \sum y_k))}.
}
