% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{trial_variance}
\alias{trial_variance}
\title{Trial variance with aggregate data}
\usage{
trial_variance(ald, tid)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{tid}{Treatment label}
}
\value{
Value
}
\description{
Calculate
\deqn{1/(\sum y_k) + 1/(N_k - \sum y_k)}.
}
