% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_outstandr.R
\name{validate_outstandr}
\alias{validate_outstandr}
\title{Input data validator}
\usage{
validate_outstandr(ipd_trial, ald_trial, strategy, CI, scale)
}
\arguments{
\item{ipd_trial}{Individual patient data}

\item{ald_trial}{Aggregate level data}

\item{strategy}{Strategy object}

\item{CI}{Confidence interval}

\item{scale}{Outcome scale}
}
\value{
No return value, called for side effects
}
\description{
Input data validator
}
\keyword{internal}
