% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{var.ratio.mcse}
\alias{var.ratio.mcse}
\title{Variability ratio MCSE}
\usage{
var.ratio.mcse(avg.se, emp.se, var.avg.se, var.emp.se)
}
\description{
Approximation of ratio variance based on independence of avg. se and emp.se
see Wolter, K., 2007. Introduction to variance estimation.
}
