% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{var.ratio}
\alias{var.ratio}
\title{Variability ratio}
\usage{
var.ratio(theta.hat, std.err)
}
\arguments{
\item{theta.hat}{Theta hat}

\item{std.err}{Standard error}
}
\value{
Ratio
}
\description{
Variability ratio
}
