% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{wald_type_interval}
\alias{wald_type_interval}
\title{Wald-type interval estimates}
\usage{
wald_type_interval(M, bar.v, b)
}
\arguments{
\item{M}{Number of syntheses used in analysis stage (high for low Monte Carlo error)}

\item{bar.v}{"within" variance (average of variance point estimates)}

\item{b}{"between" variance (sample variance of point estimates)}
}
\value{
Numeric value of Wald-type interval estimates.
}
\description{
Constructed using t-distribution with nu degrees of freedom.
}
\keyword{internal}
