% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AC_IPD}
\alias{AC_IPD}
\title{Data set for individual level patient data}
\format{
A data list including the variables needed for the mimR basic example. The variables are as follows:
\describe{
\item{list("X1")}{Numeric covariate}
\item{list("X2")}{Numeric covariate}
\item{list("X3")}{Numeric covariate}
\item{list("X4")}{Numeric covariate}
\item{list("trt")}{Integer treatment identifier}
\item{list("y")}{Integer binary outcome}
}
}
\description{
This data set contains simulated patient covariate and outcome values.
}
\references{
Remiro‐Azócar A, Heath A, Baio G (2022)
}
\keyword{datasets}
