% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALD_stats.R
\name{ALD_stats}
\alias{ALD_stats}
\title{Aggregate-level data mean and variance statistics}
\usage{
ALD_stats(strategy, ald, treatments = list("B", "C"), scale)
}
\arguments{
\item{ald}{Aggregate-level trial data}

\item{treatments}{Treatment labels list; default \code{B}, \code{C} (common; e.g. placebo)}
}
\value{
List of marginal treatment effect mean and variance
}
\description{
Aggregate-level data mean and variance statistics
}
\seealso{
\code{\link[=marginal_treatment_effect]{marginal_treatment_effect()}}, \code{\link[=marginal_variance]{marginal_variance()}}
}
