% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_ALD}
\alias{BC_ALD}
\title{Data set for aggregate level patient data}
\format{
A data list including the variables needed for the ModStanR basic example. The variables are as follows:
\describe{
\item{list("mean.X1")}{Numeric value}
\item{list("mean.X2")}{Numeric value}
\item{list("mean.X3")}{Numeric value}
\item{list("mean.X4")}{Numeric value}
\item{list("sd.X1")}{Numeric standard deviation value}
\item{list("sd.X2")}{Numeric standard deviation value}
\item{list("sd.X3")}{Numeric standard deviation value}
\item{list("sd.X4")}{Numeric standard deviation value}
\item{list("y.B.sum")}{Integer number of patients with event on B treatment}
\item{list("y.B.bar")}{Numeric proportion of patients on B treatment}
\item{list("N.B")}{Integer Total number of patients on B treatment}
\item{list("y.C.sum")}{Integer number of patients with event on C treatment}
\item{list("y.C.bar")}{Numeric proportion of patients on C treatment}
\item{list("N.C")}{Integer Total number of patients on C treatment}
}
}
\description{
This data set contains summaries of simulated patient covariate and outcome values.
}
\references{
Remiro‐Azócar A, Heath A, Baio G (2022)
}
\keyword{datasets}
