% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_bayes.R
\name{calc_gcomp_bayes}
\alias{calc_gcomp_bayes}
\title{G-computation using Stan}
\usage{
calc_gcomp_bayes(strategy, analysis_params, ...)
}
\arguments{
\item{strategy}{A list specifying the model strategy, including:
\describe{
\item{formula}{A linear regression \code{formula} object.}
\item{family}{A \code{family} object specifying the distribution and link function (e.g., \code{binomial}).}
\item{iter}{Number of iterations for the MCMC sampling.}
\item{warmup}{Number of warmup iterations for the MCMC sampling.}
\item{chains}{Number of MCMC chains.}
}}

\item{analysis_params}{List of analysis parameters}

\item{...}{Additional arguments}
}
\value{
A list of \eqn{y^*_A} and \eqn{y^*_C} posterior predictions:
\describe{
\item{\code{`0`}}{Posterior means for reference treatment group "C".}
\item{\code{`1`}}{Posterior means for comparator treatment group "A".}
}
}
\description{
Calculate draws of binary responses from posterior predictive distribution
from the Bayesian G-computation method using Hamiltonian Monte Carlo.
}
\examples{
strategy <- list(
  formula = y ~ trt:X1,
  family = binomial(),
  rho = NA,
  N = 1000L,
  trt_var = "trt",
  iter = 2000,
  warmup = 500,
  chains = 4)

ipd <- data.frame(trt = sample(c("A", "C"), size = 100, replace = TRUE),
                  X1 = rnorm(100, 1, 1),
                  y = sample(c(1,0), size = 100, prob = c(0.7,0.3), replace = TRUE))

ald <- data.frame(trt = c(NA, NA, "B", "C", "B", "C"),
                  variable = c("X1", "X1", "y", "y", NA, NA),
                  statistic = c("mean", "sd", "sum", "sum", "N", "N"),
                  value = c(0.5, 0.1, 10, 12, 20, 25))

calc_gcomp_bayes(
  strategy,
  analysis_params = 
    list(ipd = ipd, ald = ald, 
         ref_trt = "C", ipd_comp = "A"))

}
