% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_stan.R
\name{calc_gcomp_stan}
\alias{calc_gcomp_stan}
\title{G-computation using Stan}
\usage{
calc_gcomp_stan(strategy, analysis_params, ...)
}
\arguments{
\item{strategy}{A list specifying the model strategy, including:
\describe{
\item{formula}{A linear regression \code{formula} object.}
\item{family}{A \code{family} object specifying the distribution and link function (e.g., \code{binomial}).}
\item{iter}{Number of iterations for the MCMC sampling.}
\item{warmup}{Number of warmup iterations for the MCMC sampling.}
\item{chains}{Number of MCMC chains.}
}}

\item{ipd}{Individual-level patient data. Dataframe with one row per patient with outcome, treatment and covariate columns.}

\item{ald}{Aggregate-level data. Long format summary statistics for each covariate and treatment outcomes.
We assume a common distribution for each treatment arm.}
}
\value{
A list of \eqn{y^*_A} and \eqn{y^*_C} posterior predictions:
\describe{
\item{\code{`0`}}{Posterior means for reference treatment group "C".}
\item{\code{`1`}}{Posterior means for comparator treatment group "A".}
}
}
\description{
Calculate draws of binary responses from posterior predictive distribution
from the Bayesian G-computation method using Hamiltonian Monte Carlo.
}
\examples{
\dontrun{
strategy <- list(
  formula = y ~ trt + age,
  family = binomial(),
  iter = 2000,
  warmup = 500,
  chains = 4
)
ipd <- data.frame(trt = c("A", "C"),
                  y = c(1, 0),
                  age = c(30, 40))
ald <- data.frame()
calc_gcomp_stan(strategy, ipd, ald)
}
}
