% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{calc_maic}
\alias{calc_maic}
\title{Calculate MAIC}
\usage{
calc_maic(strategy, analysis_params)
}
\arguments{
\item{strategy}{Strategy object}

\item{analysis_params}{Analysis parameters; list}
}
\value{
List of numeric treatment means.
}
\description{
Calculate MAIC
}
