% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{calc_mim}
\alias{calc_mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
calc_mim(strategy, ipd, ald, ref_trt, comp_trt, ...)
}
\arguments{
\item{strategy}{Strategy object}

\item{ipd}{Individual-level patient data. Dataframe with one row per patient with outcome, treatment and covariate columns.}

\item{ald}{Aggregate-level data. Long format summary statistics for each covariate and treatment outcomes.
We assume a common distribution for each treatment arm.}

\item{ref_trt, comp_trt}{Reference and comparison treatment names}

\item{...}{Argument to pass to Stan model}
}
\description{
Multiple imputation marginalization (MIM)
}
\keyword{internal}
