% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{calc_mim}
\alias{calc_mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
calc_mim(strategy, ipd, ald, ...)
}
\arguments{
\item{strategy}{Strategy object}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{...}{Argument to pass to Stan model}
}
\description{
Multiple imputation marginalization (MIM)
}
\keyword{internal}
