% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{calc_mim}
\alias{calc_mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
calc_mim(strategy, ipd, ald, ref_trt, comp_trt, ...)
}
\arguments{
\item{strategy}{An object of class \code{strategy} created by functions such as
\code{\link[=strategy_maic]{strategy_maic()}}, \code{\link[=strategy_stc]{strategy_stc()}}, or \code{\link[=strategy_mim]{strategy_mim()}}.
Contains modelling details like the formula and family.}

\item{ipd}{Individual-level patient data. Dataframe with one row per patient with outcome, treatment and covariate columns.}

\item{ald}{Aggregate-level data. Long format summary statistics for each covariate and treatment outcomes.
We assume a common distribution for each treatment arm.}

\item{ref_trt}{Reference treatment. String.}

\item{comp_trt}{Comparator treatment. String.}

\item{...}{Additional argument to pass to Stan model}
}
\value{
Named list:
\describe{
\item{mean_comp}{Comparator mean numeric value}
\item{mean_ref}{Reference mean numeric value}
\item{hats.v}{Point estimates for the variance in each synthesis}
\item{M}{Number of posterior prediction draws}
}
}
\description{
Multiple imputation marginalization (MIM)
}
\keyword{internal}
