% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{calc_mim}
\alias{calc_mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
calc_mim(strategy, ipd, ald, ...)
}
\arguments{
\item{strategy}{Strategy object}

\item{ipd}{Individual-level patient data. Dataframe with one row per patient with outcome, treatment and covariate columns.}

\item{ald}{Aggregate-level data. Single row matrix with summary statistics for each covariate and treatment outcomes.
The format is 'mean.\emph{' and 'sd.}' for covariates and 'y.\emph{.sum', 'y.}.bar', 'y.*.sd' for treatments B and C.
We assume a common distribution for each treatment arm.}

\item{...}{Argument to pass to Stan model}
}
\description{
Multiple imputation marginalization (MIM)
}
\keyword{internal}
