% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_stc.R
\name{calc_stc}
\alias{calc_stc}
\title{Calculate simulated treatment comparison statistics}
\usage{
calc_stc(strategy, ipd, ...)
}
\value{
A list:
\describe{
\item{\code{mean_A}}{Mean for comparator treatment group "A".}
\item{\code{mean_C}}{Mean for reference treatment group "C".}
}
}
\description{
Calculate simulated treatment comparison statistics
}
