% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean}
\alias{calculate_trial_mean}
\title{Calculate Trial Mean Wrapper}
\usage{
calculate_trial_mean(ald, tid, effect, family)
}
\arguments{
\item{ald}{Aggregate level data. Data frame in long format.}

\item{tid}{Treatment ID}

\item{effect}{Effect name. String.}

\item{family}{Family distribution}
}
\value{
Numeric mean value.
}
\description{
Calculate Trial Mean Wrapper
}
