% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean_binary}
\alias{calculate_trial_mean_binary}
\title{Calculate trial mean binary}
\usage{
calculate_trial_mean_binary(ald, tid, effect)
}
\arguments{
\item{ald}{Aggregate level data}

\item{tid}{Treatment ID}

\item{effect}{Effect}
}
\value{
Value
}
\description{
Calculate trial mean binary
}
