% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean_binary}
\alias{calculate_trial_mean_binary}
\title{Calculate Trial Mean Binary Data}
\usage{
calculate_trial_mean_binary(ald, tid, effect)
}
\arguments{
\item{ald}{A Dataframe of ALD}

\item{tid}{ID}

\item{effect}{Effect}
}
\description{
Calculate Trial Mean Binary Data
}
