% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean_continuous}
\alias{calculate_trial_mean_continuous}
\title{Calculate Trial Mean Continuous Data}
\usage{
calculate_trial_mean_continuous(ald, tid, effect, verbatim = FALSE)
}
\arguments{
\item{ald}{Aggregate level data. Data frame in long format.}

\item{tid}{Treatment ID}

\item{effect}{Effect name. String.}

\item{verbatim}{Print messages, logical}
}
\value{
Numeric mean value.
}
\description{
Calculate Trial Mean Continuous Data
}
