% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_mean_count}
\alias{calculate_trial_mean_count}
\title{Calculate Trial Mean Count Data}
\usage{
calculate_trial_mean_count(ald, tid, effect)
}
\arguments{
\item{ald}{Aggregate level data. Data frame in long format.}

\item{tid}{Treatment ID}

\item{effect}{Effect name. String.}
}
\value{
Numeric mean value.
}
\description{
Calculate Trial Mean Count Data
}
