% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{calculate_trial_variance_count}
\alias{calculate_trial_variance_count}
\title{Calculate trial variance count}
\usage{
calculate_trial_variance_count(ald, tid, effect)
}
\arguments{
\item{ald}{Aggregate level data. Data frame in long format.}

\item{tid}{Treatment ID.}

\item{effect}{Effect name. String.}
}
\value{
Numeric value of total variance.
}
\description{
Calculate trial variance count
}
