% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{continuity_correction}
\alias{continuity_correction}
\title{Continuity Correction}
\usage{
continuity_correction(ald, correction = 0.5)
}
\arguments{
\item{ald}{A dataframe of aggregate level data}

\item{correction}{Continuity correction numeric size. Default to 0.5.}
}
\value{
Corrected aggregate level data
}
\description{
Continuity Correction
}
\keyword{internal}
