% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{continuity_correction}
\alias{continuity_correction}
\title{Continuity Correction}
\usage{
continuity_correction(ald, correction = 0.5, verbatim = FALSE)
}
\arguments{
\item{ald}{Aggregate level data. Data frame in long format.}

\item{correction}{Continuity correction numeric size. Default to 0.5.}

\item{verbatim}{Print messages, logical}
}
\value{
Corrected aggregate level data frame.
}
\description{
Continuity Correction
}
\keyword{internal}
