% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var_sandwich.R
\name{estimate_var_sandwich}
\alias{estimate_var_sandwich}
\title{Estimate Variance Sandwich Estimator}
\usage{
estimate_var_sandwich(strategy, ipd_data, ...)
}
\arguments{
\item{strategy}{An object of class \code{strategy} created by functions such as
\code{\link[=strategy_maic]{strategy_maic()}}, \code{\link[=strategy_stc]{strategy_stc()}}, or \code{\link[=strategy_mim]{strategy_mim()}}.
Contains modelling details like the formula and family.}

\item{ipd_data}{IPD data}

\item{...}{Additional arguments}
}
\value{
Numeric variance estimate
}
\description{
Estimate Variance Sandwich Estimator
}
