% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_var_sandwich.R
\name{estimate_var_sandwich}
\alias{estimate_var_sandwich}
\title{Estimate Variance Sandwich Estimator}
\usage{
estimate_var_sandwich(strategy, ipd_data, ...)
}
\arguments{
\item{strategy}{Strategy}

\item{ipd_data}{IPD data}

\item{...}{Additional arguments}
}
\value{
Numeric variance estimate
}
\description{
Estimate Variance Sandwich Estimator
}
