% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml.boot}
\alias{gcomp_ml.boot}
\title{G-computation maximum likelihood bootstrap}
\usage{
gcomp_ml.boot(
  data,
  indices,
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2")
)
}
\arguments{
\item{data}{Trial data}

\item{indices}{Indices sampled from rows of \code{data}}

\item{formula}{Linear regression formula; default \eqn{y = X_3 + X_4 + \beta_t X_1 + \beta_t X_2"}}
}
\value{
Mean difference in expected log-odds
}
\description{
Using bootstrap resampling, calculates the log odds ratio
}
