% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml.boot}
\alias{gcomp_ml.boot}
\title{G-computation maximum likelihood bootstrap}
\usage{
gcomp_ml.boot(
  data,
  indices,
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2")
)
}
\value{
mean difference in expected log-odds
}
\description{
G-computation maximum likelihood bootstrap
}
