% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml.boot}
\alias{gcomp_ml.boot}
\title{G-computation maximum likelihood bootstrap}
\usage{
gcomp_ml.boot(
  data,
  indices,
  R,
  formula = NULL,
  family,
  trt_var,
  ref_trt = NA,
  comp_trt = NA,
  rho = NA,
  N = 1000,
  ald
)
}
\arguments{
\item{data}{IPD trial data}

\item{indices}{Indices sampled from rows of \code{data} for bootstrapping}

\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt + trt:X2. For covariates as both PF and EM use \code{*} syntax.}

\item{family}{A 'family' object specifying the distribution and link function (e.g., 'binomial').
See stats::family() for more details.}

\item{rho}{A named square matrix of covariate correlations; default NA.}

\item{N}{Synthetic sample size for g-computation}

\item{ald}{Aggregate-level data for covariates.}
}
\value{
Relative treatment effect
}
\description{
Using bootstrap resampling, calculates the relative treatment effect,
such as log odds ratio, log relative risk or risk difference.
}
\examples{
\dontrun{
data <- data.frame(trt = c("A", "C"),
                   y = c(1, 0))
gcomp_ml.boot(data, indices = 1:2, formula = y ~ trt,
              R = 100, family = binomial(), N = 1000, ald = NULL)
}
}
\seealso{
\code{\link[=strategy_gcomp_ml]{strategy_gcomp_ml()}}, \code{\link[=gcomp_ml_log_odds_ratio]{gcomp_ml_log_odds_ratio()}}
}
\keyword{internal}
