% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml_log_odds_ratio}
\alias{gcomp_ml_log_odds_ratio}
\title{marginal A vs. C log-odds ratio (mean difference in expected log-odds)
estimated by transforming from probability to linear predictor scale}
\usage{
gcomp_ml_log_odds_ratio(formula, dat)
}
\description{
marginal A vs. C log-odds ratio (mean difference in expected log-odds)
estimated by transforming from probability to linear predictor scale
}
