% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{gcomp_ml_log_odds_ratio}
\alias{gcomp_ml_log_odds_ratio}
\title{G-computation Maximum Likelihood Log-Odds Ratio}
\usage{
gcomp_ml_log_odds_ratio(formula, dat)
}
\arguments{
\item{formula}{Linear regression \code{formula} object}

\item{dat}{Trial data}
}
\value{
Difference of log-odds
}
\description{
Marginal \emph{A} vs \emph{C} log-odds ratio (mean difference in expected log-odds)
estimated by transforming from probability to linear predictor scale.
}
\details{
\eqn{\log(\hat{\mu}_A/(1 - \hat{\mu}_A)) - \log(\hat{\mu}_C/(1 - \hat{\mu}_C))}
}
\seealso{
\code{\link[=strategy_gcomp_ml]{strategy_gcomp_ml()}}, \code{\link[=gcomp_ml.boot]{gcomp_ml.boot()}}
}
