% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_stan.R
\name{gcomp_stan}
\alias{gcomp_stan}
\title{G-computation using Stan}
\usage{
gcomp_stan(formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"), ipd, ald)
}
\arguments{
\item{formula}{Linear regression formula object}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}
}
\value{
A list of \code{y.star.A} and \code{y.star.C} posterior predictions
}
\description{
G-computation using Stan
}
