% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_stan.R
\name{gcomp_stan}
\alias{gcomp_stan}
\title{gcomp_stan}
\usage{
gcomp_stan(formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"), dat = AC.IPD)
}
\arguments{
\item{formula}{}

\item{dat}{}
}
\description{
gcomp_stan
}
