% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulated datasets for index and comparator trials}
\usage{
gen_data(N, b_trt, b_X, b_EM, b_0, meanX, sdX, event_rate, corX, allocation)
}
\arguments{
\item{N}{Total number of patients}

\item{b_trt}{\code{b} coefficient for active treatment vs. common comparator}

\item{b_X}{\code{b} coefficients for each prognostic variable \code{X}}

\item{b_EM}{\code{b} coefficients effect modifiers}

\item{b_0}{Intercept coefficient}

\item{meanX}{Mean of each normally-distributed covariate \code{X}}

\item{sdX}{Standard deviation of each covariate \code{X}}

\item{event_rate}{Event rate}

\item{corX}{Covariate correlation coefficient of \code{X}}

\item{allocation}{Allocation to active treatment as proportion of total; 0 to 1}
}
\value{
Data frame of \code{X}, \code{trt} and \code{y}
}
\description{
Generate simulated datasets for index and comparator trials
}
\examples{

gen_data(
 N = 100,
 b_trt = log(0.17),
 b_X = -log(0.5),
 b_EM = -log(0.67),
 b_0 = ,
 meanX = 0.6,
 sdX = 0.4,
 event_rate = 0.35, 
 corX = 0.2,
 allocation = 2/3) 

}
\keyword{internal}
