% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulated datasets for index and comparator trials}
\usage{
gen_data(N, b_trt, b_X, b_EM, b_0, meanX, sdX, event_rate, corX, allocation)
}
\arguments{
\item{N}{Size}

\item{b_trt}{b coefficient for treatment}

\item{b_X}{b coefficients for X}

\item{b_EM}{b coefficients effect modifiers}

\item{b_0}{Intercept coefficient}

\item{meanX}{mean X}

\item{sdX}{standard deviation X}

\item{event_rate}{Event rate}

\item{corX}{Correlation of X}

\item{allocation}{Allocation}
}
\value{
Data frame of \code{X}, \code{trt} and \code{y}
}
\description{
Generate simulated datasets for index and comparator trials
}
\keyword{internal}
