% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulated datasets for index and comparator trials}
\usage{
gen_data(N, b_trt, b_X, b_EM, b_0, meanX, sdX, event_rate, corX, allocation)
}
\arguments{
\item{N}{N}

\item{b_trt}{b treatment}

\item{b_X}{b_X}

\item{b_EM}{b_EM}

\item{b_0}{b_0}

\item{meanX}{mean X}

\item{sdX}{sd X}

\item{event_rate}{Event rate}

\item{corX}{Correlation of X}

\item{allocation}{Allocation}
}
\value{
Data frame of \code{X}, \code{trt} and \code{y}
}
\description{
Generate simulated datasets for index and comparator trials
}
