% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_covariate_names}
\alias{get_covariate_names}
\title{Get covariate names}
\usage{
get_covariate_names(formula)
}
\arguments{
\item{formula}{Linear regression formula}
}
\value{
covariate names
}
\description{
Get covariate names
}
