% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_eff_mod_names}
\alias{get_eff_mod_names}
\title{Get effect modifiers}
\usage{
get_eff_mod_names(formula, trt_var)
}
\arguments{
\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt + trt:X2. For covariates as both PF and EM use \code{*} syntax.}
}
\value{
Effect modifiers names
}
\description{
Get effect modifiers
}
\keyword{internal}
