% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_mean_names}
\alias{get_mean_names}
\title{Get mean names}
\usage{
get_mean_names(ald, keep_nms)
}
\arguments{
\item{ald}{Aggregate-level data. Single row matrix with summary statistics for each covariate and treatment outcomes.
The format is 'mean.\emph{' and 'sd.}' for covariates and 'y.\emph{.sum', 'y.}.bar', 'y.*.sd' for treatments B and C.
We assume a common distribution for each treatment arm.}

\item{keep_nms}{Variable names character vector}
}
\value{
Mean names vector
}
\description{
Get mean names
}
\keyword{internal}
