% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_mean_names}
\alias{get_mean_names}
\title{Get mean names}
\usage{
get_mean_names(dat, var_names)
}
\arguments{
\item{dat}{}

\item{var_names}{}
}
\value{
mean names
}
\description{
Get mean names
}
