% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{get_ref_trt}
\alias{get_ref_trt}
\title{Get reference treatment}
\usage{
get_ref_trt(ref_trt, trt, ipd_trial, ald_trial)
}
\arguments{
\item{ref_trt}{Reference treatment}

\item{trt}{Treatment}

\item{ipd_trial}{A dataframe of IPD}

\item{ald_trial}{A dataframe of ALD}
}
\value{
String of reference treatment name.
}
\description{
Get reference treatment
}
\keyword{internal}
