% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_sd_names}
\alias{get_sd_names}
\title{Get SD names}
\usage{
get_sd_names(dat, var_names)
}
\arguments{
\item{dat}{}

\item{var_names}{}
}
\value{
SD names
}
\description{
Get SD names
}
\keyword{internal}
