% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_sd_names}
\alias{get_sd_names}
\title{Get standard deviation names}
\usage{
get_sd_names(ald, var_names)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{var_names}{Variable names character vector}
}
\value{
Standard deviation names vector
}
\description{
Get standard deviation names
}
\keyword{internal}
