% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ate.R
\name{get_treatment_effect}
\alias{get_treatment_effect}
\title{Get treatment effect scale corresponding to a link function}
\usage{
get_treatment_effect(link)
}
\description{
Get treatment effect scale corresponding to a link function
}
