% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_treatment_name}
\alias{get_treatment_name}
\title{Get treatment name}
\usage{
get_treatment_name(formula)
}
\arguments{
\item{formula}{Linear regression \code{formula} object. Prognostic factors (PF) are main effects and effect modifiers (EM) are
interactions with the treatment variable, i.e., y ~ X1 + trt:X2. For covariates as both PF and EM use \code{*} syntax.}
}
\value{
Treatment name
}
\description{
Get treatment name
}
\keyword{internal}
