% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{get_treatment_name}
\alias{get_treatment_name}
\title{Get treatment name}
\usage{
get_treatment_name(formula)
}
\arguments{
\item{formula}{}
}
\value{
treatment name
}
\description{
Get treatment name
}
