% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{hat_Delta_stats}
\alias{hat_Delta_stats}
\title{Main wrapper}
\usage{
hat_Delta_stats(AC.IPD, BC.ALD, strategy, ...)
}
\arguments{
\item{AC.IPD}{}

\item{BC.ALD}{}

\item{strategy}{}

\item{...}{Additional arguments}
}
\description{
Main wrapper
}
