% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats}
\alias{IPD_stats}
\alias{IPD_stats.default}
\alias{IPD_stats.mim}
\title{Calculate individual-level patient data statistics}
\usage{
IPD_stats(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{default}(...)

\method{IPD_stats}{mim}(strategy, ipd, ald, scale, ...)
}
\arguments{
\item{strategy}{A list corresponding to different approaches}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{...}{Additional arguments}
}
\value{
Mean and variance values
}
\description{
Separate methods for each approach
MAIC, STC, G-computation via MLE or Bayesian inference.
}
\section{Multiple imputation marginalisation}{

Using Stan, compute marginal relative treatment effect for \emph{A} vs \emph{C} for each MCMC sample
by transforming from probability to linear predictor scale. Approximate by
using imputation and combining estimates using Rubin's rules, in contrast to \code{\link[=IPD_stats.gcomp_stan]{IPD_stats.gcomp_stan()}}.
}

