% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{IPD_stats}
\alias{IPD_stats}
\alias{IPD_stats.default}
\alias{IPD_stats.mim}
\alias{IPD_stats.stc}
\alias{IPD_stats.maic}
\alias{IPD_stats.gcomp_ml}
\alias{IPD_stats.gcomp_stan}
\title{Calculate individual-level patient data statistics}
\usage{
IPD_stats(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{default}(...)

\method{IPD_stats}{mim}(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{stc}(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{maic}(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{gcomp_ml}(strategy, ipd, ald, scale, ...)

\method{IPD_stats}{gcomp_stan}(strategy, ipd, ald, scale, ...)
}
\arguments{
\item{strategy}{A list corresponding to different approaches}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{scale}{A scaling parameter for the effect calculation.}

\item{...}{Additional arguments}
}
\value{
A list containing:
\describe{
\item{mean}{Estimated mean treatment effect.}
\item{var}{Estimated variance of the treatment effect.}
}
}
\description{
Computes mean and variance statistics for individual-level patient data using various approaches,
including Matching-Adjusted Indirect Comparison (MAIC), Simulated Treatment Comparison (STC),
and G-computation via Maximum Likelihood Estimation (MLE) or Bayesian inference.
}
\section{Multiple imputation marginalisation}{

Using Stan, compute marginal relative treatment effect for \emph{A} vs \emph{C} for each MCMC sample
by transforming from probability to linear predictor scale. Approximate by
using imputation and combining estimates using Rubin's rules, in contrast to \code{\link[=IPD_stats.gcomp_stan]{IPD_stats.gcomp_stan()}}.
}

\section{Simulated treatment comparison statistics}{

IPD from the \emph{AC} trial are used to fit a regression model describing the
observed outcomes \eqn{y} in terms of the relevant baseline characteristics \eqn{x} and
the treatment variable \eqn{z}.
}

\section{Matching-adjusted indirect comparison statistics}{

Marginal \emph{A} vs \emph{C} treatment effect estimates
using bootstrapping sampling.
}

\section{G-computation maximum likelihood statistics}{

Compute a non-parametric bootstrap with default \eqn{R=1000} resamples.
}

\section{G-computation Bayesian statistics}{

Using Stan, compute marginal log-odds ratio for \emph{A} vs \emph{C} for each MCMC sample
by transforming from probability to linear predictor scale.
}

\examples{
\dontrun{
strategy <- strategy_maic()
ipd <- data.frame(id = 1:100, treatment = sample(c("A", "C"), 100, replace = TRUE), outcome = rnorm(100))
ald <- data.frame(treatment = c("A", "C"), mean = c(0.2, 0.1), var = c(0.05, 0.03))
IPD_stats(strategy, ipd, ald, scale = "log")
}
}
