% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic.boot}
\alias{maic.boot}
\title{MAIC bootstrap sample}
\usage{
maic.boot(
  data,
  indices,
  balance_matrix,
  outcome_x_matrix,
  outcome_y,
  ald_targets,
  scaling_factors,
  trt_var,
  family,
  hat_w = NULL,
  ipd = NULL,
  outcome_model = NULL,
  balance_model = NULL,
  ald = NULL,
  moments = 1,
  int = FALSE
)
}
\arguments{
\item{data}{Individual-level patient data (data frame).}

\item{indices}{Vector of indices, same length as original,
which define the bootstrap sample.}

\item{balance_matrix}{Pre-computed balance matrix.}

\item{outcome_x_matrix}{Pre-computed outcome design matrix.}

\item{outcome_y}{Pre-computed outcome vector.}

\item{ald_targets}{Vector of ALD targets.}

\item{scaling_factors}{Vector of scaling factors.}

\item{trt_var}{Treatment variable name.}

\item{family}{A 'family' object specifying the distribution and link function.}

\item{hat_w}{MAIC weights; default \code{NULL} which calls \code{\link[=maic_weights]{maic_weights()}}.}

\item{ipd}{Backwards compatibility IPD data (optional).}

\item{outcome_model}{Backwards compatibility outcome model formula (optional).}

\item{balance_model}{Backwards compatibility balance model formula (optional).}

\item{ald}{Backwards compatibility ALD data (optional).}

\item{moments}{Backwards compatibility moments (default 1).}

\item{int}{Backwards compatibility interactions flag (default FALSE).}
}
\value{
Vector of fitted probabilities for treatments \emph{A} and \emph{C}
}
\description{
Matching-adjusted indirect comparison bootstrap sampling.
}
\seealso{
\code{\link[=calc_IPD_stats.maic]{calc_IPD_stats.maic()}}
}
\keyword{internal}
