% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic.boot}
\alias{maic.boot}
\title{MAIC bootstrap}
\usage{
maic.boot(ipd, indices, formula, ald)
}
\arguments{
\item{ipd}{Individual-level data}

\item{indices}{Vector of indices, same length as original,
which define the bootstrap sample}

\item{formula}{Linear regression formula}

\item{ald}{Aggregate-level data}
}
\value{
Fitted treatment coefficient is marginal effect for \emph{A} vs \emph{C}
}
\description{
MAIC bootstrap
}
