% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic.boot}
\alias{maic.boot}
\title{MAIC bootstrap}
\usage{
maic.boot(data, indices, formula, dat_ALD)
}
\arguments{
\item{data}{Original data}

\item{indices}{Vector of indices, same length as original,
which define the bootstrap sample}

\item{formula}{Linear regression formula}

\item{dat_ALD}{Aggregate-level data}
}
\value{
Fitted treatment coefficient is marginal effect for A vs C
}
\description{
MAIC bootstrap
}
