% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic.boot}
\alias{maic.boot}
\title{MAIC bootstrap sample}
\usage{
maic.boot(ipd, indices = 1:nrow(ipd), formula, family, ald, hat_w = NULL)
}
\arguments{
\item{ipd}{Individual-level data}

\item{indices}{Vector of indices, same length as original,
which define the bootstrap sample}

\item{formula}{Linear regression formula}

\item{family}{A character string specifying the family distribution (e.g., "binomial").}

\item{ald}{Aggregate-level data}

\item{hat_w}{MAIC weights; default \code{NULL} which calls \code{\link[=maic_weights]{maic_weights()}}}
}
\value{
Vector of fitted probabilities for treatments \emph{A} and \emph{C}
}
\description{
Matching-adjusted indirect comparison bootstrap sampling.
}
\seealso{
\code{\link[=IPD_stats.maic]{IPD_stats.maic()}}
}
\keyword{internal}
