% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{maic_weights}
\alias{maic_weights}
\title{Estimate MAIC weights}
\usage{
maic_weights(X.EM)
}
\arguments{
\item{X.EM}{Centred S=1 effect modifiers; matrix or data frame}
}
\value{
Estimated weights for each individual; vector
}
\description{
Estimate MAIC weights
}
\references{
J. E. Signorovitch et al,
Comparative Effectiveness Without Head-to-Head Trials:
A Method for Matching-Adjusted Indirect Comparisons Applied
to Psoriasis Treatment with Adalimumab or Etanercept,
Pharmacoeconomics 2010; 28 (10): 935-945
}
