% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{marginal_treatment_effect}
\alias{marginal_treatment_effect}
\title{B vs C marginal treatment effect from reported event counts}
\usage{
marginal_treatment_effect(x, trials = list("B", "C"))
}
\description{
\leqn{\log(n_B n_{\bar{C}})  -log(n_C n_{\bar{B}})}
}
