% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALD_stats.R
\name{marginal_variance}
\alias{marginal_variance}
\title{Marginal effect variance using the delta method}
\usage{
marginal_variance(ald, treatments = list("B", "C"), family)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{treatments}{Treatment labels list; default \emph{B} vs \emph{C}}
}
\value{
Sum of variances
}
\description{
Calculate
\deqn{\frac{1}{n_C} + \frac{1}{n_{\bar{C}}} + \frac{1}{n_B} + \frac{1}{n_{\bar{B}}}}.
}
