% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{marginal_variance}
\alias{marginal_variance}
\title{marginal effect variance using the delta method}
\usage{
marginal_variance(x, trials = list("B", "C"))
}
\description{
\eqn{\frac{1}{n_C} + \frac{1}{n_{\bar{C}}} + \frac{1}{n_B} + \frac{1}{n_{\bar{B}}}}
}
