% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ALD_stats.R
\name{marginal_variance}
\alias{marginal_variance}
\title{Marginal effect variance using the delta method}
\usage{
marginal_variance(ald, treatments = list("B", "C"), scale, family)
}
\arguments{
\item{ald}{Aggregate-level data}

\item{treatments}{A list of treatment labels; default \emph{B} vs \emph{C}}

\item{scale}{A scaling parameter for the calculation.}

\item{family}{A character string specifying the family distribution (e.g., "binomial").}
}
\value{
The total variance of marginal treatment effects.
}
\description{
Computes the total variance of marginal treatment effects using the delta method.
}
\examples{
\dontrun{
ald <- data.frame(trt = c("B","C","B","C"),
                  variable = c(NA, NA, "y", "y"),
                  statistic = c("N", "N", "sum", "sum"),
                  value = c(100, 100, 50, 60)
marginal_variance(ald, treatments = list("B", "C"), scale = "log", family = "binomial")
}
}
