% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{mim}
\alias{mim}
\title{Multiple imputation marginalization (MIM)}
\usage{
mim(ipd.index, ipd.target, M = 1000, n.chains = 2, warmup = 1000, iters = 4000)
}
\arguments{
\item{ipd.index, ipd.target}{Index RCT and target covariate datasets}

\item{M}{Number of syntheses used in analysis stage (high for low Monte Carlo error)}
}
\description{
Multiple imputation marginalization (MIM)
}
