% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.outstandR}
\alias{plot.outstandR}
\title{Default Plot Method for outstandR Objects}
\usage{
\method{plot}{outstandR}(x, ..., type = c("both", "contrasts", "absolute"), labels = NULL)
}
\arguments{
\item{x}{An object of class 'outstandR' or a list of 'outstandR' objects.}

\item{...}{Additional 'outstandR' objects for comparison.}

\item{type}{Character, one of "both" (default), "contrasts", or "absolute".}

\item{labels}{Optional character vector of names for the models.}
}
\description{
Default Plot Method for outstandR Objects
}
