% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_ald}
\alias{prep_ald}
\title{Prepare Aggregate Level Data}
\usage{
prep_ald(form, data, trt_var = "trt")
}
\arguments{
\item{form}{Formula}

\item{data}{A dataframe of aggregate level data}

\item{trt_var}{Treatment variable name}
}
\value{
A data frame of filtered ALD by variables in formula.
}
\description{
Prepare Aggregate Level Data
}
\keyword{internal}
