% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{Q}
\alias{Q}
\title{Objective function to minimize for standard method of moments MAIC}
\usage{
Q(beta, X)
}
\arguments{
\item{beta}{Beta}

\item{X}{X}
}
\description{
Objective function to minimize for standard method of moments MAIC
}
\keyword{internal}
