% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{reshape_ald_to_wide}
\alias{reshape_ald_to_wide}
\title{Convert from long to wide format}
\usage{
reshape_ald_to_wide(df)
}
\description{
Convert from long to wide format
}
\examples{
df <- 
  data.frame(
    variable = c("age", "age", "y", "y", "y", "y", "y", "y", "y", "y"),
    statistic = c("mean", "sd", "sum", "bar", "sd", "N", "sum", "bar", "sd", "N"),
    treatment = c(NA, NA, "B", "B", "B", "B", "C", "C", "C", "C"),
    value = c(1,1,1,1,1,1,1,1,1,1))

}
