% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_stats.R
\name{result_stats}
\alias{result_stats}
\title{Result Statistics}
\usage{
result_stats(AC_stats, BC_stats, CI = 0.95)
}
\arguments{
\item{AC_stats, BC_stats}{}

\item{CI}{Confidence interval 1-alpha}
}
\value{
List
}
\description{
Result Statistics
}
