% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_stats.R
\name{result_stats}
\alias{result_stats}
\title{Result Statistics}
\usage{
result_stats(ipd_stats, ald_stats, CI = 0.95)
}
\arguments{
\item{ipd_stats, ald_stats}{}

\item{CI}{Confidence interval 1-alpha}
}
\value{
List
}
\description{
Combining output from aggregate level data studies BC and
adjusted individual level data studies AC into a single object.
}
\keyword{internal}
