% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp_ml.R
\name{simulate_ALD_pseudo_pop}
\alias{simulate_ALD_pseudo_pop}
\title{Synthetic cohort using normal copula}
\usage{
simulate_ALD_pseudo_pop(formula, ipd, ald, N = 1000)
}
\arguments{
\item{formula}{Linear regression \code{formula} object}

\item{ipd}{Individual-level data}

\item{ald}{Aggregate-level data}

\item{N}{Sample size}
}
\description{
Synthetic cohort using normal copula
}
\keyword{internal}
