% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{strategy_maic}
\alias{strategy_maic}
\alias{strategy_stc}
\alias{strategy_gcomp_ml}
\alias{strategy_gcomp_stan}
\alias{strategy}
\alias{new_strategy}
\title{New strategy objects}
\usage{
strategy_maic(
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"),
  R = 1000,
  dat_ALD = BC.ALD
)

strategy_stc(
  formula = as.formula("y ~ X3 + X4 + trt*I(X1 - mean(X1)) + trt*I(X2 - mean(X2))")
)

strategy_gcomp_ml(
  formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"),
  R = 1000
)

strategy_gcomp_stan(formula = as.formula("y ~ X3 + X4 + trt*X1 + trt*X2"))

new_strategy(strategy, ...)
}
\arguments{
\item{formula}{Linear regression formula object}

\item{R}{The number of resamples used for the non-parametric bootstrap}

\item{strategy}{Class name}

\item{...}{Additional arguments}

\item{ald}{Aggregate-level data}
}
\value{
\code{maic} class object

\code{stc} class object

\code{gcomp_ml} class object

\code{gcomp_stan} class object
}
\description{
Create class for each approach
}
\section{Matching-adjusted indirect comparison (MAIC)}{

Used to compare marginal
treatment effects where there are cross-trial differences in effect modifiers
and limited patient-level data.

The default formula is
\deqn{
 y = X3 + X4 + trt*X1 + trt*X2
}
}

\section{Simulated treatment comparison (STC)}{

Outcome regression-based method which targets a conditional treatment effect.

The default formula is
\deqn{
 y = X3 + X4 + trt*(X1 - mean(X1)) + trt*(X2 - mean(X2))
}
}

\section{G-computation maximum likelihood}{


The default formula is
\deqn{
 y = X3 + X4 + trt*X1 + trt*X2
}
}

\section{G-computation Bayesian}{


The default formula is
\deqn{
 y = X3 + X4 + trt*X1 + trt*X2
}
}

