% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPD_stats.R
\name{strategy_stc}
\alias{strategy_stc}
\title{strategy_stc}
\usage{
strategy_stc(
  formula =
    as.formula("y ~ X3 + X4 +\\n                                   trt*I(X1 - mean(X1)) +\\n                                   trt*I(X2 - mean(X2))")
)
}
\arguments{
\item{formula}{}
}
\description{
strategy_stc
}
