% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{trial_treatment_effect}
\alias{trial_treatment_effect}
\title{Trial treatment effect}
\usage{
trial_treatment_effect(x, k)
}
\arguments{
\item{x}{x}

\item{k}{k}
}
\value{
Value
}
\description{
Trial treatment effect
}
