% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_outstandr.R
\name{validate_outstandr}
\alias{validate_outstandr}
\title{Input data validator}
\usage{
validate_outstandr(ipd_trial, ald_trial, strategy, CI, scale)
}
\description{
Input data validator
}
\keyword{internal}
