% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_outstandr.R
\name{validate_outstandr}
\alias{validate_outstandr}
\title{Input data validator}
\usage{
validate_outstandr(ipd_trial, ald_trial, strategy, CI, scale)
}
\arguments{
\item{ipd_trial}{Individual patient data}

\item{ald_trial}{Aggregate level data}

\item{strategy}{An object of class \code{strategy} created by functions such as
\code{\link[=strategy_maic]{strategy_maic()}}, \code{\link[=strategy_stc]{strategy_stc()}}, or \code{\link[=strategy_mim]{strategy_mim()}}.
Contains modelling details like the formula and family.}

\item{CI}{Confidence interval}

\item{scale}{Outcome scale}
}
\value{
No return value, called for side effects
}
\description{
Input data validator
}
\keyword{internal}
