% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{var.ratio.mcse}
\alias{var.ratio.mcse}
\title{Variability ratio MCSE}
\usage{
var.ratio.mcse(avg.se, emp.se, var.avg.se, var.emp.se)
}
\arguments{
\item{avg.se}{Average SE}

\item{emp.se}{Emp SE}

\item{var.avg.se}{Variance of average SE}

\item{var.emp.se}{Variance of Emp SE}
}
\description{
Approximation of ratio variance based on independence of avg. se and emp.se
see \insertCite{wolter2007}{mimR}
}
\references{
\insertRef{wolter2007}{mimR}
}
\keyword{internal}
