% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{var_by_pooling}
\alias{var_by_pooling}
\title{Variance estimate by pooling}
\usage{
var_by_pooling(M, bar.v, b)
}
\arguments{
\item{M}{Number of syntheses used in analysis stage (high for low Monte Carlo error)}

\item{bar.v}{"within" variance (average of variance point estimates)}

\item{b}{"between" variance (sample variance of point estimates)}
}
\description{
Use combining rules to estimate
}
\keyword{internal}
