% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{var_by_pooling}
\alias{var_by_pooling}
\title{Variance estimate by pooling
Use combining rules to estimate}
\usage{
var_by_pooling(M, bar.v, b)
}
\description{
Variance estimate by pooling
Use combining rules to estimate
}
\keyword{internal}
