% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mim.R
\name{wald_type_interval}
\alias{wald_type_interval}
\title{Wald-type interval estimates
Constructed using t-distribution with nu degrees of freedom}
\usage{
wald_type_interval(M, bar.v, b)
}
\description{
Wald-type interval estimates
Constructed using t-distribution with nu degrees of freedom
}
\keyword{internal}
